fixed default value for frame pixel format widget, missing in cb1361633203080a33589b6f4b6f2fc0af0ebdc6: "switched default frame pixel format format from 32-bit float to 16-bit float (half)".

+++ src/appleseed/renderer/modeling/frame/frame.cpp
@@ -770,7 +770,7 @@ DictionaryArray FrameFactory::get_widget_definitions()
                     .insert("Floating-Point, 32-bit", "float")
                     .insert("Floating-Point, 64-bit", "double"))
             .insert("use", "optional")
-            .insert("default", "float"));
+            .insert("default", "half"));
 
     definitions.push_back(
         Dictionary()
